apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gim-backend"
version = "0.1.0"
description = "FastAPI backend with ML service for IssueIndex"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.9",
    
    # Auth & Security
    "ua-parser>=1.0.0",
    "cryptography>=43.0.0",
    
    # HTTP Client
    "httpx>=0.28.1",
    
    # Configuration
    "pydantic-settings>=2.7.0",
    
    # GCP Cloud Tasks
    "google-cloud-tasks>=2.16.0",

    # GCP Pub/Sub
    "google-cloud-pubsub>=2.0.0",
    
    # Note: Install gim-database first with: pip install -e ../../packages/database
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "testcontainers[postgres]>=4.9.0",
    "fakeredis>=2.26.0",
    "greenlet>=3.0.0",
    "psycopg2-binary>=2.9.9",
]
cpu = [
    # CPU-only PyTorch for memory efficiency
    "torch>=2.5.1",
]
ml = [
    # Local ML embeddings (alternative to Vertex AI)
    "sentence-transformers>=3.3.1",
    "numpy>=2.2.1",
]
resume = [
    # Resume parsing dependencies
    "docling>=2.0.0",
    "gliner>=0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["gim_backend"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]
